requirements.txt
numpy
PyYAML
xlrd
PyQt5
matplotlib